requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2025.9.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.19.0
onnx-ir==0.1.9
onnxruntime==1.22.1
onnxscript==0.5.2
onnxslim==0.1.69
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
protobuf==6.32.1
psutil==7.1.0
pybind11==3.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
six==1.17.0
sympy==1.14.0
tifffile==2025.9.20
torch==2.8.0
torchvision==0.23.0
triton==3.4.0
typing_extensions==4.15.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
